Improved elapsed time column #141
The elapsed time between a log entry and the start of the log file is displayed up to the milisecond. The width of the column depends on the range of values in the data source, but currently only the number of days is omitted if the range is less than a day.