Fixed building nupckg with newer nuget toolset nuget.exe contains a bug where if one specifies the .csproj file, then it will spout nonsense errors about a supposedly missing dependency group:
"Error NU5128: Some target frameworks declared in the dependencies group of the nuspec and the lib/ref folder do not have exact matches in the other location. Consult the list of actions below:
- Add a dependency group for .NETFramework4.5 to the nuspec"

However this is nothing but a giant waste of time and can be circumvented by specifying the nuspec file **only**.

Also moved the call to create the nuget package to a new batch file so that the appveyor jobs won't have to be modified that often.