Trying to update nuspec to nugets ridiculous standards (#247)
nuget.exe contains a bug where if one specifies the .csproj file, then it will spout nonsense errors about a supposedly missing dependency group:

Error NU5128: Some target frameworks declared in the dependencies group of the nuspec and the lib/ref folder do not have exact matches in the other location. Consult the list of actions below:

Add a dependency group for .NETFramework4.5 to the nuspec
Adding said group to the nuspec file did not change anything and the same error occured regardless.

The solution was to not specify the csproj file but to only specify the nuspec file, adding the entries which were previously added automatically from the csproj myself.

The the process, I moved the call to create the nuget package to a new batch file so that the appveyor jobs won't have to be modified that often.