Fixed bug where highlighted search terms were misaligned when tabs were involved #244 (#246)
The central text viewer would misalign the search-term highlights incorrectly for every line containing tabs. This was caused by the search itself operating on the lines as is and reference sections of the line with respect to the number of characters from the start, e.g. a result would be index 12, 5 characters long.
However the text display itself would modify the message prior to display and effectively expand tab characters by the number of spaces the user has specified. However when applying the search result, it would not take this expansion into account and re-use the indices reported by the search *on* the expanded message.
This means that as soon as the user specified a tab width > 1, the search results were misaligned by the number of tabs *left* of the search term, multiplied by the number spaces per tab.

The fix involved simply subdividing the message into its search result constituents first and only then expanding the individual substrings, replacing tabs with spaces.