Build scripts create portable archive (#359)
Every Tailviewer build (both mainline and MRs) creates a "portable" version which is merely a zipped version of the installation folder under "%PROGRAMFILES%\Tailviewer". This version should be executable directly from a USB drive since it doesn't write to the installation directory.

However there is still the major drawback that the computer needs to have .NET 4.7.1 installed, which is why this isn't a truly portable version. That will be addressed once Tailviewer has been ported to >= .NET 5.